requirements.txt
asgiref==3.5.2
astroid==2.12.12
autopep8==1.7.0
black==22.10.0
click==8.1.3
dill==0.3.5.1
Django==3.0
django-taggit==1.2.0
docutils==0.19
flake8==5.0.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
pylint==2.15.5
pytz==2022.5
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
wrapt==1.14.1
